requirements.txt
colorama==0.4.6
speedtest-cli==2.1.3
pyinstaller==6.10.0

